requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
asttokens==3.0.0
attrs==24.2.0
bidict==0.23.1
boto3==1.35.72
botocore==1.35.72
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
docutils==0.20.1
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.5
frozenlist==1.5.0
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
icecream==2.1.3
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.22
langchain-groq==0.2.1
langchain-openai==0.2.11
langsmith==0.1.147
markdown2==2.5.1
MarkupSafe==3.0.2
multidict==6.1.0
nicegui==2.7.0
nicegui-highcharts==2.0.2
numpy==2.1.3
openai==1.57.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pendulum==3.0.0
pillow==11.0.0
propcache==0.2.1
pscript==0.7.7
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.19
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
time-machine==2.16.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.3
